+++ assets/js/script.v120.js
@@ -1,5 +1,5 @@
 /**
- * TelePrompter - Open Source TelePrompter
+ * TelePrompter v1.2.0 - Browser-based TelePrompter with Remote Control
  * (c) 2021 Peter Schmalfeldt
  * License: https://github.com/manifestinteractive/teleprompter/blob/master/LICENSE
  */
@@ -14,7 +14,8 @@ var TelePrompter = (function() {
   var $elm = {};
 
   /* App Settings */
-  var debug = true,
+  var emitTimeout,
+    debug = false,
     initialized = false,
     isPlaying = false,
     remote,
@@ -22,7 +23,11 @@ var TelePrompter = (function() {
     socket,
     modalOpen = false,
     timeout,
-    timer;
+    timer,
+    timerExp = 10,
+    timerGA,
+    timerURL,
+    version = 'v1.2.0';
 
   /* Default App Settings */
   var defaultConfig = {
@@ -61,31 +66,61 @@ var TelePrompter = (function() {
     $elm.buttonReset = $('.button.reset');
     $elm.closeModal = $('.close-modal');
     $elm.fontSize = $('.font_size');
+    $elm.gaInput = $('input[data-ga], textarea[data-ga], select[data-ga]');
+    $elm.gaLinks = $('a[data-ga], button[data-ga]');
     $elm.header = $('header');
     $elm.headerContent = $('header h1, header nav');
     $elm.markerOverlay = $('.marker, .overlay');
+    $elm.modal = $('#modal');
     $elm.remoteID = $('.remote-id');
-    $elm.remoteModal = $('.remote-modal');
     $elm.remoteURL = $('.remote-url');
+    $elm.remoteControlModal = $('#remote-control-modal');
     $elm.speed = $('.speed');
+    $elm.softwareUpdate = $('#software-update');
     $elm.teleprompter = $('#teleprompter');
     $elm.textColor = $('#text-color');
     $elm.window = $(window);
 
     // Bind Events
-    $elm.backgroundColor.on('change', handleBackgroundColor);
-    $elm.buttonDimControls.on('click', handleDim);
-    $elm.buttonFlipX.on('click', handleFlipX);
-    $elm.buttonFlipY.on('click', handleFlipY);
-    $elm.buttonPlay.on('click', handlePlay);
-    $elm.buttonRemote.on('click', handleRemote);
-    $elm.buttonReset.on('click', handleReset);
-    $elm.closeModal.on('click', handleCloseModal);
-    $elm.textColor.on('change', handleTextColor);
+    $elm.backgroundColor.on('change.teleprompter', handleBackgroundColor);
+    $elm.buttonDimControls.on('click.teleprompter', handleDim);
+    $elm.buttonFlipX.on('click.teleprompter', handleFlipX);
+    $elm.buttonFlipY.on('click.teleprompter', handleFlipY);
+    $elm.buttonPlay.on('click.teleprompter', handlePlay);
+    $elm.buttonRemote.on('click.teleprompter', handleRemote);
+    $elm.buttonReset.on('click.teleprompter', handleReset);
+    $elm.closeModal.on('click.teleprompter', handleCloseModal);
+    $elm.gaInput.on('change.teleprompter', gaInput);
+    $elm.gaLinks.on('click.teleprompter', gaLinks);
+    $elm.textColor.on('change.teleprompter', handleTextColor);
 
     // Listen for Key Presses
     $elm.teleprompter.keyup(updateTeleprompter);
     $elm.body.keydown(navigate);
+  }
+
+  /**
+   * Check for Software Update
+   */
+   function checkForUpdate() {
+    var currentVersion = localStorage.getItem('teleprompter_version');
+    var customText = localStorage.getItem('teleprompter_text');
+
+    // Compare User Config to Default Config
+    if (customText && currentVersion !== version) {
+      // User had custom settings on page load
+      localStorage.setItem('teleprompter_version', version);
+
+      // Open Software Update Modal
+      $elm.modal.css('display', 'flex');
+      $elm.remoteControlModal.hide();
+      $elm.softwareUpdate.show();
+
+      modalOpen = true;
+
+    } else if (currentVersion !== version) {
+      localStorage.setItem('teleprompter_version', version);
+    }
   }
 
   /**
@@ -94,7 +129,6 @@ var TelePrompter = (function() {
   function init() {
     // Exit if already started
     if (initialized) {
-      console.warn('TelePrompter Already Initialized');
       return;
     }
 
@@ -103,6 +137,7 @@ var TelePrompter = (function() {
     initSettings();
     initUI();
     initRemote();
+    checkForUpdate();
 
     // Track that we've started TelePrompter
     initialized = true;
@@ -128,6 +163,10 @@ var TelePrompter = (function() {
     if (debug) {
       console.log('[TP]', 'Remote Initialized', currentRemote ? '( Remote ID: ' + currentRemote + ' )' : '( No Remote )');
     }
+
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){}, timerExp);
+    gaEvent('TP', 'Remote Initialized', currentRemote ? currentRemote : 'No Remote');
   }
 
   /**
@@ -264,6 +303,10 @@ var TelePrompter = (function() {
     if (debug) {
       console.log('[TP]', 'Settings Initialized', urlParams ? urlParams : '( No URL Params )');
     }
+
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){}, timerExp);
+    gaEvent('TP', 'Settings Initialized', urlParams ? 'Custom URL Params' : 'No URL Params');
   }
 
   /**
@@ -379,6 +422,63 @@ var TelePrompter = (function() {
   }
 
   /**
+   * Setup Events using Google Analytics
+   * @param category
+   * @param action
+   * @param label
+   * @param value
+   */
+   function gaEvent(category, action, label, value) {
+    if (typeof gtag !== 'undefined') {
+      gtag('event', action, {
+        event_category: category,
+        event_label: label,
+        value: value
+      });
+    }
+
+    if (debug) {
+      console.log('[GA]', category, action, label, value);
+    }
+  }
+
+  /**
+   * Setup Google Analytics on Links
+   * @param event
+   */
+  function gaLinks(event){
+    var data;
+
+    if (typeof event.target !== 'undefined' && typeof event.target.dataset !== 'undefined' && typeof event.target.dataset.ga !== 'undefined') {
+      data = event.target.dataset;
+    } else if (typeof event.target !== 'undefined' && typeof event.target.parentNode !== 'undefined' && typeof event.target.parentNode.dataset !== 'undefined' && typeof event.target.parentNode.dataset.track !== 'undefined') {
+      data = event.target.parentNode.dataset;
+    }
+
+    if (typeof data === 'object' && typeof data.category === 'string' && typeof data.action === 'string' && typeof data.label === 'string') {
+      gaEvent(data.category, data.action, data.label, data.value);
+    }
+  }
+
+  /**
+   * Setup Google Analytics on Input
+   * @param event
+   */
+  function gaInput(event){
+    var data;
+
+    if (typeof event.target !== 'undefined' && typeof event.target.dataset !== 'undefined' && typeof event.target.dataset.ga !== 'undefined') {
+      data = event.target.dataset;
+    } else if (typeof event.target !== 'undefined' && typeof event.target.parentNode !== 'undefined' && typeof event.target.parentNode.dataset !== 'undefined' && typeof event.target.parentNode.dataset.track !== 'undefined') {
+      data = event.target.parentNode.dataset;
+    }
+
+    if (typeof data === 'object' && typeof data.category === 'string' && typeof data.action === 'string') {
+      gaEvent(data.category, data.action, event.target.value, event.target.value.length);
+    }
+  }
+
+  /**
    * Get App Config
    * @param {String} key
    * @returns Object
@@ -418,14 +518,23 @@ var TelePrompter = (function() {
     localStorage.setItem('teleprompter_background_color', config.backgroundColor);
 
     if (socket && remote) {
-      socket.emit('clientCommand', 'updateConfig', config);
+      clearTimeout(emitTimeout);
+      emitTimeout = setTimeout(function(){
+        socket.emit('clientCommand', 'updateConfig', config);
+      }, timerExp);
     }
 
     if (debug) {
       console.log('[TP]', 'Background Color Changed:', config.backgroundColor);
     }
 
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){}, timerExp);
+    gaEvent('TP', 'Background Color Changed', config.backgroundColor);
+
     // Update URL Params
+    clearTimeout(timerURL);
+    timerURL = setTimeout(updateURL, timerExp);
     updateURL();
   }
 
@@ -433,8 +542,14 @@ var TelePrompter = (function() {
    * Handle Closing Modal
    */
   function handleCloseModal() {
-    $elm.remoteModal.hide();
-    $elm.buttonRemote.focus();
+    // Reset Focus on Remote Button if needed
+    if ($elm.remoteControlModal.is(':visible')) {
+      $elm.buttonRemote.focus();
+    }
+
+    $elm.modal.hide();
+    $elm.remoteControlModal.hide();
+    $elm.softwareUpdate.hide();
 
     modalOpen = false;
   }
@@ -463,15 +578,24 @@ var TelePrompter = (function() {
     localStorage.setItem('teleprompter_dim_controls', config.dimControls);
 
     if (socket && remote && !skipUpdate) {
-      socket.emit('clientCommand', 'updateConfig', config);
+      clearTimeout(emitTimeout);
+      emitTimeout = setTimeout(function(){
+        socket.emit('clientCommand', 'updateConfig', config);
+      }, timerExp);
     }
 
     if (debug) {
       console.log('[TP]', 'Dim Control Changed:', config.dimControls);
     }
 
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){
+      gaEvent('TP', 'Dim Control Changed', config.dimControls);
+    }, timerExp);
+
     // Update URL Params
-    updateURL();
+    clearTimeout(timerURL);
+    timerURL = setTimeout(updateURL, timerExp);
   }
 
   /**
@@ -508,14 +632,24 @@ var TelePrompter = (function() {
     localStorage.setItem('teleprompter_flip_x', config.flipX);
 
     if (socket && remote && !skipUpdate) {
-      socket.emit('clientCommand', 'updateConfig', config);
+      clearTimeout(emitTimeout);
+      emitTimeout = setTimeout(function(){
+        socket.emit('clientCommand', 'updateConfig', config);
+      }, timerExp);
     }
 
     if (debug) {
       console.log('[TP]', 'Flip X Changed:', config.flipX);
     }
 
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){
+      gaEvent('TP', 'Flip X Changed', config.flipX);
+    }, timerExp);
+
     // Update URL Params
+    clearTimeout(timerURL);
+    timerURL = setTimeout(updateURL, timerExp);
     updateURL();
   }
 
@@ -567,14 +701,24 @@ var TelePrompter = (function() {
     }
 
     if (socket && remote && !skipUpdate) {
-      socket.emit('clientCommand', 'updateConfig', config);
+      clearTimeout(emitTimeout);
+      emitTimeout = setTimeout(function(){
+        socket.emit('clientCommand', 'updateConfig', config);
+      }, timerExp);
     }
 
     if (debug) {
       console.log('[TP]', 'Flip Y Changed:', config.flipY);
     }
 
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){
+      gaEvent('TP', 'Flip Y Changed', config.flipY);
+    }, timerExp);
+
     // Update URL Params
+    clearTimeout(timerURL);
+    timerURL = setTimeout(updateURL, timerExp);
     updateURL();
   }
 
@@ -588,14 +732,24 @@ var TelePrompter = (function() {
     localStorage.setItem('teleprompter_text_color', config.textColor);
 
     if (socket && remote) {
-      socket.emit('clientCommand', 'updateConfig', config);
+      clearTimeout(emitTimeout);
+      emitTimeout = setTimeout(function(){
+        socket.emit('clientCommand', 'updateConfig', config);
+      }, timerExp);
     }
 
     if (debug) {
       console.log('[TP]', 'Text Color Changed:', config.textColor);
     }
 
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){
+      gaEvent('TP', 'Text Color Changed', config.textColor);
+    }, timerExp);
+
     // Update URL Params
+    clearTimeout(timerURL);
+    timerURL = setTimeout(updateURL, timerExp);
     updateURL();
   }
 
@@ -618,7 +772,9 @@ var TelePrompter = (function() {
       var currentRemote = localStorage.getItem('teleprompter_remote_id');
       remoteConnect(currentRemote);
     } else {
-      $elm.remoteModal.css('display', 'flex');
+      $elm.modal.css('display', 'flex');
+      $elm.remoteControlModal.show();
+      $elm.softwareUpdate.hide();
     }
 
     $elm.buttonRemote.blur();
@@ -646,7 +802,10 @@ var TelePrompter = (function() {
 
     if (socket && remote) {
       socket.emit('clientCommand', 'updateTime', '00:00:00');
-      socket.emit('clientCommand', 'updateConfig', config);
+      clearTimeout(emitTimeout);
+      emitTimeout = setTimeout(function(){
+        socket.emit('clientCommand', 'updateConfig', config);
+      }, timerExp);
     }
 
     if (debug) {
@@ -682,8 +841,11 @@ var TelePrompter = (function() {
 
     // Check if Escape Key and Modal Open
     if (evt.keyCode == escape && modalOpen) {
-      $elm.remoteModal.hide();
-      $elm.buttonRemote.focus();
+      if ($elm.remoteControlModal.is(':visible')) {
+        $elm.buttonRemote.focus();
+      }
+
+      $elm.modal.hide();
       modalOpen = false;
       evt.preventDefault();
       evt.stopPropagation();
@@ -805,7 +967,10 @@ var TelePrompter = (function() {
       config.pageScrollPercent = Math.round(((scrollTop / (scrollHeight - clientHeight)) + Number.EPSILON) * 100);
 
       if (socket && remote) {
-        socket.emit('clientCommand', 'updateConfig', config);
+        clearTimeout(emitTimeout);
+        emitTimeout = setTimeout(function(){
+          socket.emit('clientCommand', 'updateConfig', config);
+        }, timerExp);
       }
     }, animate);
   }
@@ -863,12 +1028,17 @@ var TelePrompter = (function() {
       $elm.remoteID.text(remote);
 
       if (!currentRemote) {
-        $elm.remoteModal.css('display', 'flex');
+        $elm.modal.css('display', 'flex');
       }
 
       if (debug) {
         console.log('[IO]', 'Socket Connected');
       }
+
+      clearTimeout(timerGA);
+      timerGA = setTimeout(function(){
+        gaEvent('IO', 'Socket Connected');
+      }, timerExp);
     });
 
     socket.on('disconnect', function() {
@@ -878,22 +1048,41 @@ var TelePrompter = (function() {
       if (debug) {
         console.log('[IO]', 'Socket Disconnected');
       }
+
+      clearTimeout(timerGA);
+      timerGA = setTimeout(function(){
+        gaEvent('IO', 'Socket Disconnected');
+      }, timerExp);
     });
 
     socket.on('connectedToRemote', function() {
       localStorage.setItem('teleprompter_remote_id', remote);
       $elm.buttonRemote.addClass('active');
-      socket.emit('clientCommand', 'updateConfig', config);
+
+      clearTimeout(emitTimeout);
+      emitTimeout = setTimeout(function(){
+        socket.emit('clientCommand', 'updateConfig', config);
+      }, timerExp);
 
       if (debug) {
         console.log('[IO]', 'Remote Connected:', remote);
       }
+
+      clearTimeout(timerGA);
+      timerGA = setTimeout(function(){
+        gaEvent('IO', 'Remote Connected', remote);
+      }, timerExp);
     });
 
     socket.on('remoteControl', function(command, value) {
       if (debug) {
         console.log('[TP]', 'remoteControl', command, value);
       }
+
+      clearTimeout(timerGA);
+      timerGA = setTimeout(function(){
+        gaEvent('IO', 'Remote Control', command);
+      }, timerExp);
 
       switch (command) {
         case 'reset':
@@ -909,16 +1098,20 @@ var TelePrompter = (function() {
           break;
 
         case 'hideModal':
-          $elm.remoteModal.hide();
+          $elm.modal.hide();
           break;
 
         case 'getConfig':
           if (socket && remote) {
-            socket.emit('clientCommand', 'updateConfig', config);
+            clearTimeout(emitTimeout);
+            emitTimeout = setTimeout(function(){
+              socket.emit('clientCommand', 'updateConfig', config);
+            }, timerExp);
           }
           break;
 
         case 'updateConfig':
+          clearTimeout(emitTimeout);
           remoteUpdate(config, value);
           break;
       }
@@ -937,6 +1130,11 @@ var TelePrompter = (function() {
     if (debug) {
       console.log('[IO]', 'Remote Disconnected');
     }
+
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){
+      gaEvent('IO', 'Remote Disconnected');
+    }, timerExp);
   }
 
   /**
@@ -950,6 +1148,11 @@ var TelePrompter = (function() {
       console.log('[IO]', 'Old Config:', oldConfig);
       console.log('[IO]', 'New Config:', newConfig);
     }
+
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){
+      gaEvent('IO', 'Remote Update');
+    }, timerExp);
 
     if (oldConfig.dimControls !== newConfig.dimControls) {
       handleDim(null, true);
@@ -994,6 +1197,8 @@ var TelePrompter = (function() {
     }
 
     // Update URL Params
+    clearTimeout(timerURL);
+    timerURL = setTimeout(updateURL, timerExp);
     updateURL();
   }
 
@@ -1003,7 +1208,6 @@ var TelePrompter = (function() {
   function startTeleprompter() {
     // Check if Already Playing
     if (isPlaying) {
-      console.warn('[TP]', 'Starting TelePrompter', '( Already Started )');
       return;
     }
 
@@ -1029,6 +1233,11 @@ var TelePrompter = (function() {
     if (debug) {
       console.log('[TP]', 'Starting TelePrompter');
     }
+
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){
+      gaEvent('TP', 'Starting TelePrompter');
+    }, timerExp);
   }
 
   /**
@@ -1037,7 +1246,6 @@ var TelePrompter = (function() {
   function stopTeleprompter() {
     // Check if Already Stopped
     if (!isPlaying) {
-      console.warn('[TP]', 'Stopping TelePrompter', '( Not Running )');
       return;
     }
 
@@ -1063,6 +1271,11 @@ var TelePrompter = (function() {
     if (debug) {
       console.log('[TP]', 'Stopping TelePrompter');
     }
+
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){
+      gaEvent('TP', 'Stopping TelePrompter');
+    }, timerExp);
   }
 
   /**
@@ -1091,14 +1304,24 @@ var TelePrompter = (function() {
     }
 
     if (socket && remote && !skipUpdate) {
-      socket.emit('clientCommand', 'updateConfig', config);
+      clearTimeout(emitTimeout);
+      emitTimeout = setTimeout(function(){
+        socket.emit('clientCommand', 'updateConfig', config);
+      }, timerExp);
     }
 
     if (debug) {
       console.log('[TP]', 'Font Size Changed:', config.fontSize);
     }
 
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){
+      gaEvent('TP', 'Font Size Changed', config.fontSize);
+    }, timerExp);
+
     // Update URL Params
+    clearTimeout(timerURL);
+    timerURL = setTimeout(updateURL, timerExp);
     updateURL();
   }
 
@@ -1116,14 +1339,24 @@ var TelePrompter = (function() {
     }
 
     if (socket && remote && !skipUpdate) {
-      socket.emit('clientCommand', 'updateConfig', config);
+      clearTimeout(emitTimeout);
+      emitTimeout = setTimeout(function(){
+        socket.emit('clientCommand', 'updateConfig', config);
+      }, timerExp);
     }
 
     if (debug) {
       console.log('[TP]', 'Page Speed Changed:', config.pageSpeed);
     }
 
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){
+      gaEvent('TP', 'Page Speed Changed', config.pageSpeed);
+    }, timerExp);
+
     // Update URL Params
+    clearTimeout(timerURL);
+    timerURL = setTimeout(updateURL, timerExp);
     updateURL();
   }
 
@@ -1146,6 +1379,11 @@ var TelePrompter = (function() {
     if (debug) {
       console.log('[TP]', 'TelePrompter Text Updated');
     }
+
+    clearTimeout(timerGA);
+    timerGA = setTimeout(function(){
+      gaEvent('TP', 'TelePrompter Text Updated');
+    }, timerExp);
   }
 
   /**
@@ -1176,7 +1414,7 @@ var TelePrompter = (function() {
 
   /* Expose Select Control to Public TelePrompter Object */
   return {
-    version: 'v1.2.0',
+    version: version,
     init: init,
     getConfig: getConfig,
     start: startTeleprompter,
@@ -1184,59 +1422,32 @@ var TelePrompter = (function() {
     reset: handleReset,
     setDebug: function(bool) {
       debug = !!bool;
-
-      console.log('debug', debug)
-
       return this;
     },
     setSpeed: function(speed) {
       speed = Math.min(50, Math.max(0, speed));
       $elm.speed.slider('value', parseInt(speed));
-
       return this;
     },
     setFontSize: function(size) {
       size = Math.min(100, Math.max(12, size));
       $elm.fontSize.slider('value', parseInt(size));
-
       return this;
     },
     setDim: function(bool) {
       config.dimControls = !bool;
       handleDim();
-
       return this;
     },
     setFlipX: function(bool) {
       config.flipX = !bool;
       handleFlipX();
-
       return this;
     },
     setFlipY: function(bool) {
       config.flipY = !bool;
       handleFlipY();
-
       return this;
     }
   }
-
-  // FIXED: Pull URL Params on Page Load and Update UI & Local Storage
-  // TODO: Add Version Update Modal to show Changes to Existing / New Users ( track with version number in Local Storage )
-  // TODO: Update Favicon and Apple Icons for Teleprompter ( but leave Remotes as is )
-  // TODO: Address Responsive Layout Issues
-  // TODO: Listen for Manual Page Scroll and update Config
-  // FIXED: Remove Unused Code
-  // TODO: Perform Accessibility Tests for UI, Keyboard Navigation & Screen Readers
-  // FIXED: Update Third Party Libraries ( if easy )
-  // TODO: Add ability to run Website as an App in supported browsers
-  // TODO: Optimize CSS & JS ( remove lingering console statements )
-  // TODO: Make sure pasting from Word still works like it used to
-  // FIXED: Add Changelog to Repo
-  // FIXED: Look into why refreshing the Teleprompter Disconnects the Remote ( that should not happen )
 })();
-
-// Handle Onload Event
-window.onload = function() {
-  TelePrompter.init();
-};